# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BookBridge – Student Book Exchange Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get the current user's wishlist
      responses:
        "200":
          description: Wishlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WishlistItem"

    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add a book to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistInput"
      responses:
        "201":
          description: Added to wishlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishlistItem"
  /wishlist/{bookId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove a book from wishlist
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Exchange Requests
  /exchanges:
    get:
      operationId: listExchanges
      tags: [exchanges]
      summary: List exchange requests for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Exchange requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExchangeRequest"

    post:
      operationId: createExchange
      tags: [exchanges]
      summary: Send an exchange request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeInput"
      responses:
        "201":
          description: Created exchange request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRequest"
  /exchanges/{id}:
    patch:
      operationId: updateExchange
      tags: [exchanges]
      summary: Accept, reject, or complete an exchange request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeUpdate"
      responses:
        "200":
          description: Updated exchange request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRequest"

  # Student Dashboard
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Book:
      type: object
      required: [id, title, author, publisher, category, department, semester, condition, price, listingType, isAvailable, sellerId, sellerName, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        category:
          type: string
        department:
          type: string
        semester:
          type: string
        condition:
          type: string
        price:
          type: number
        listingType:
          type: string
          enum: [sell, exchange, donate]
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        sellerId:
          type: integer
        sellerName:
          type: string
        viewCount:
          type: integer
        wishlistCount:
          type: integer
        createdAt:
          type: string
...
    WishlistItem:
      type: object
      required: [id, bookId, userId, book, addedAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        book:
          $ref: "#/components/schemas/Book"
        addedAt:
          type: string
    WishlistInput:
      type: object
      required: [bookId]
      properties:
        bookId:
          type: integer

    # Exchanges
    ExchangeRequest:
      type: object
      required: [id, bookId, requesterId, ownerId, status, createdAt]
      properties:
        id:
          type: integer
        bookId:
          type: integer
        requesterId:
          type: integer
        ownerId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected, completed]
        message:
          type: ["string", "null"]
        book:
          $ref: "#/components/schemas/Book"
        requesterName:
          type: string
        requesterEmail:
          type: string
        ownerName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    ExchangeInput:
      type: object
      required: [bookId]
      properties:
        bookId:
          type: integer
        message:
          type: string
    ExchangeUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, rejected, completed]

    # Dashboard
...